requirements.txt
torch~=2.0.1
torchvision~=0.15.2
numpy~=1.21.5
scikit-learn~=1.0.2
tqdm~=4.64.1
opencv-python~=4.8.0.74
matplotlib~=3.5.2
pillow~=9.2.0